// apps/web/src/app/api/model/task.ts
/**
 * Talos API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TaskPriority } from './taskPriority';
import { TaskRiskLevel } from './taskRiskLevel';
import { TaskStatus } from './taskStatus';


export interface Task { 
    id: string;
    projectId: string;
    title: string;
    description?: string;
    source: string;
    status: TaskStatus;
    priority: TaskPriority;
    riskLevel: TaskRiskLevel;
    boardPosition: number;
    requestedBy?: string;
    assignedAgentKey?: string;
    createdAt: string;
    updatedAt: string;
}
export namespace Task {
}