// apps/web/src/app/api/model/taskStatus.ts
/**
 * Talos API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const TaskStatus = {
    BACKLOG: 'BACKLOG',
    READY: 'READY',
    RUNNING: 'RUNNING',
    REVIEW: 'REVIEW',
    BLOCKED: 'BLOCKED',
    DONE: 'DONE',
    CANCELLED: 'CANCELLED'
} as const;
export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];